+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Add support for setting the `host` splunk tag.  [#24](https://github.com/splunk/splunk-library-javalogging/issues/24).
 * Send additional logger attributes (logger name, thread name, context/property maps) to the event. [](#)
+* Add support for exception reporting as its own field
+* Add support for marker including as its own field
 
 ## Version 1.5.1